refactor: extract reusable SuggestionList component from field and function suggestions
