+++ packages/frontend/src/components/common/SuggestionList/index.ts
@@ -1,0 +1,5 @@
+export { generateSuggestion } from './generateSuggestion';
+export {
+    generateFieldSuggestion,
+    type FieldSuggestionItem,
+} from './generateFieldSuggestion';
